tests/wpt/tests/css/css-transitions/parsing/transition-delay-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Transitions: parsing transition-delay with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-transitions/#propdef-transition-delay">
<meta name="assert" content="transition-delay supports the full grammar '<time> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("transition-delay", '0s');
test_valid_value("transition-delay", '500ms');
test_valid_value("transition-delay", '1s, 2s');
test_valid_value("transition-delay", '-1s, -2s');
</script>
</body>
</html>
